Implemented ANR and Crash Tests in Cart screen

+++ app/src/main/java/com/bluetriangle/bluetriangledemo/ui/cart/CartViewModel.kt
@@ -1,5 +1,6 @@
 package com.bluetriangle.bluetriangledemo.ui.cart
 
+import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -39,16 +40,30 @@ class CartViewModel @Inject constructor(val cartRepository: CartRepository) : Vi
         }
     }
 
-    fun removeCartItem(cartItem: CartItem) {
+    fun removeCartItem(cartItem: CartItem, startTime: Long = System.currentTimeMillis()) {
+        val diff = System.currentTimeMillis() - startTime
+        Log.d("CartViewModelTag", "removeCartItem: $diff")
+        if(diff <= 10_000) {
+            Thread.sleep(200)
+            removeCartItem(cartItem, startTime)
+        }
+    }
+
+    fun reduceQuantity(cartItem: CartItem) {
         viewModelScope.launch(Dispatchers.IO) {
-            try {
-                val cart = cartRepository.removeCartItem(cartItem)
-                withContext(Dispatchers.Main) {
-                    _cart.value = cart
-                }
-            } catch (e: Exception) {
-                errorHandler.showError(e)
+            if(cartItem.quantity <= 1) {
+                cartRepository.removeCartItem(cartItem)
+            } else {
+                cartRepository.reduceQuantity(cartItem)
             }
+            refreshCart()
+        }
+    }
+
+    fun increaseQuantity(cartItem: CartItem) {
+        viewModelScope.launch(Dispatchers.IO) {
+            cartRepository.increaseQuantity(cartItem)
+            refreshCart()
         }
     }
 }
